# backend/contracts/04-api-contract.openapi.yaml
openapi: 3.0.3
info:
  title: ERP/CRM Backend API
  version: 0.1.0
  description: |
    Контракт backend API для первого этапа внедрения NestJS backend-слоя.

    Backend является источником истины для авторизации, прав доступа, транзакционного
    сохранения заказов, финансовых расчётов, экспорта, VLM-интеграций и аудита.

    Frontend не должен напрямую управлять транзакциями заказа, refresh-token'ами,
    финансовыми итогами, экспортными payload'ами и проверкой прав доступа.

    Stage-1 backend API использует явное path versioning. Текущая стабильная
    версия: /api/v1. Неломающие поля можно добавлять в v1; breaking changes
    требуют /api/v2 и отдельного cutover/deprecation плана.
servers:
  - url: http://localhost:3000
    description: Local NestJS backend
  - url: https://api.example.com
    description: Production backend placeholder
...
paths:
...
  /api/v1/orders/{orderId}:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    get:
      tags:
        - Orders
      operationId: getOrderById
      summary: Получить заказ
      description: Возвращает полный заказ для show/edit формы.
      security:
        - bearerAuth: []
      x-permission: orders.view
      responses:
        '200':
          description: Полный заказ
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - Orders
      operationId: updateOrder
      summary: Обновить заказ
      description: |
        Обновляет заказ и все дочерние сущности в одной БД-транзакции. Обязательна
        проверка optimistic locking через поле version.
      security:
        - bearerAuth: []
      x-permission: orders.update
      x-transactional: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveOrderDto'
      responses:
        '200':
          description: Заказ обновлён
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveOrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/OrderVersionConflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Orders
      operationId: deleteOrder
      summary: Удалить или архивировать заказ
      description: |
        Выполняет soft-delete/archive заказа, а не физическое удаление.
        Команда stale-safe: клиент обязан передать текущую версию заказа в
        `If-Match`, а retry-safe поведение обеспечивается `Idempotency-Key`.
      security:
        - bearerAuth: []
      x-permission: orders.delete
      x-transactional: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
      responses:
        '200':
          description: Заказ удалён/архивирован
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteOrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          description: Версия заказа устарела или idempotency key уже используется другим запросом
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
              examples:
                versionConflict:
                  value:
                    error:
                      code: ORDER_VERSION_CONFLICT
                      message: Заказ был изменён другим пользователем
                      details:
                        currentVersion: 8
                        clientVersion: 7
                      requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
                idempotencyKeyReused:
                  value:
                    error:
                      code: IDEMPOTENCY_KEY_REUSED
                      message: Idempotency key was reused with a different request
                      details:
                        idempotencyKey: order-delete:01HZX8K5T3M2F5F7K9A2S1QW3E
                      requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
                idempotencyInProgress:
                  value:
                    error:
                      code: IDEMPOTENCY_IN_PROGRESS
                      message: Idempotent command is still processing
                      details:
                        idempotencyKey: order-delete:01HZX8K5T3M2F5F7K9A2S1QW3E
                      requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
                idempotencyFailed:
                  value:
                    error:
                      code: IDEMPOTENCY_FAILED
                      message: Idempotent command previously failed
                      details:
                        idempotencyKey: order-delete:01HZX8K5T3M2F5F7K9A2S1QW3E
                      requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{orderId}/status:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    patch:
      tags:
        - Production Actions
      operationId: changeOrderStatus
      summary: Изменить статус заказа
      description: Изменяет статус заказа с записью audit event и stale-safe проверкой version.
      security:
        - bearerAuth: []
      x-permissions:
        - orders.change_status
        - orders.update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeOrderStatusRequest'
      responses:
        '200':
          description: Статус изменён
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionActionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/OrderVersionConflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/FeatureDisabled'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{orderId}/payment-status:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    patch:
      tags:
        - Production Actions
      operationId: changePaymentStatus
      summary: Изменить статус оплаты заказа
      description: Изменяет ручной статус оплаты заказа с записью audit event и stale-safe проверкой version.
      security:
        - bearerAuth: []
      x-permissions:
        - orders.update
        - payments.update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePaymentStatusRequest'
      responses:
        '200':
          description: Статус оплаты изменён
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionActionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/OrderVersionConflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/FeatureDisabled'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{orderId}/production-status:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    patch:
      tags:
        - Production Actions
      operationId: changeProductionStatus
      summary: Изменить текущий производственный статус заказа
      description: Изменяет текущий производственный статус заказа с записью audit event и stale-safe проверкой version.
      security:
        - bearerAuth: []
      x-permissions:
        - orders.change_production_status
        - orders.update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeProductionStatusRequest'
      responses:
        '200':
          description: Changed order production current status
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionActionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/OrderVersionConflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/FeatureDisabled'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{orderId}/calendar-date:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    patch:
      tags:
        - Production Actions
      operationId: moveOrderCalendarDate
      summary: Перенести плановую дату заказа
      security:
        - bearerAuth: []
      x-permissions:
        - calendar.view
        - orders.update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveCalendarDateRequest'
      responses:
        '200':
          description: Плановая дата изменена
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionActionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/OrderVersionConflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/FeatureDisabled'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{orderId}/order-status:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    patch:
      tags:
        - Production Actions
      operationId: changeOrderStatusLegacy
      summary: Изменить статус заказа через legacy route
      security:
        - bearerAuth: []
      x-permissions:
        - orders.change_status
        - orders.update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeOrderStatusRequest'
      responses:
        '200':
          description: Статус изменён
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionActionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/OrderVersionConflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/FeatureDisabled'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token из /api/v1/auth/login или /api/v1/auth/refresh.
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly refresh token cookie. Не доступен frontend JavaScript-коду.
  headers:
    RequestId:
      description: Уникальный id запроса для логов и аудита.
      schema:
        type: string
        example: req_01HZX8K5T3M2F5F7K9A2S1QW3E
  parameters:
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    IfMatchVersion:
      name: If-Match
      in: header
      required: true
      description: Текущая версия заказа для optimistic locking. Допускается число или quoted ETag вида `"4"`.
      schema:
        type: string
        example: '"4"'
...
  responses:
...
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          examples:
            authRequired:
              value:
                error:
                  code: AUTH_REQUIRED
                  message: Требуется авторизация
                  requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          examples:
            permissionDenied:
              value:
                error:
                  code: PERMISSION_DENIED
                  message: Недостаточно прав для выполнения действия
                  details:
                    requiredPermission: orders.update
                  requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
    OrderNotFound:
      description: Заказ не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          examples:
            notFound:
              value:
                error:
                  code: ORDER_NOT_FOUND
                  message: Заказ не найден
                  requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
...
    OrderVersionConflict:
      description: Конфликт версии заказа
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          examples:
            versionConflict:
              value:
                error:
                  code: ORDER_VERSION_CONFLICT
                  message: Заказ был изменён другим пользователем
                  details:
                    currentVersion: 8
                    clientVersion: 7
                  requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
...
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          examples:
            internal:
              value:
                error:
                  code: INTERNAL_ERROR
                  message: Внутренняя ошибка сервера
                  requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
  schemas:
...
    OrderResponse:
      type: object
      required:
        - order
      properties:
        order:
          $ref: '#/components/schemas/OrderDto'
    SaveOrderResponse:
      type: object
      required:
        - order
      properties:
        order:
          $ref: '#/components/schemas/OrderDto'
...